Completely reworked the TP system under the hood
This should allow me to continue further development on the mod with minimal spagetti, and less so struggling to remember my arbitrary fixed point system

+++ Content/ITensionConsumingItem.cs
@@ -8,7 +8,7 @@ using Terraria;
 namespace Delterra.Content {
     public interface ITensionConsumingItem {
 
-        public abstract int GetBaseTPCost(Player player);
+        public abstract double GetBaseTPCost(Player player);
 
         public virtual bool IsTPConsumedOnUse(Player player) {
             return true;
